Attempt at only updating database if updating is needed (failed attempt).
I'm making this commit public in case anyone wants to give this a shot.

I basically tried when a player leaves, we only save their selections to the database if current selections are different from what's stored in the database.

If someone can help figure this out, great! If not, no worries!